Better editors in grid (meal, table, drink), fixed adding new element to grid

+++ PresentationLayer/BookYourFood/Models/TableViewModel.cs
@@ -2,7 +2,7 @@ namespace BookYourFood.Models
 {
     public class TableViewModel
     {
-        public long TableNumber { get; set; }
+        public int TableNumber { get; set; }
         public int SeatsNumber { get; set; }
         public long Id { get; set; }
     }
